refactor: rename new Button component
