+++ src/frontend/src/eth/components/tokens/AddTokenReview.svelte
@@ -6,8 +6,8 @@
 	import { infuraErc20Providers } from '$eth/providers/infura-erc20.providers';
 	import type { Erc20Metadata } from '$eth/types/erc20';
 	import AddTokenWarning from '$lib/components/tokens/AddTokenWarning.svelte';
+	import Button from '$lib/components/ui/Button.svelte';
 	import ButtonGroup from '$lib/components/ui/ButtonGroup.svelte';
-	import ButtonPrimary from '$lib/components/ui/ButtonPrimary.svelte';
 	import ContentWithToolbar from '$lib/components/ui/ContentWithToolbar.svelte';
 	import TextWithLogo from '$lib/components/ui/TextWithLogo.svelte';
 	import Value from '$lib/components/ui/Value.svelte';
@@ -139,8 +139,8 @@
 		<button class="secondary block flex-1" on:click={() => dispatch('icBack')}
 			>{$i18n.core.text.back}</button
 		>
-		<ButtonPrimary disabled={invalid} on:click={() => dispatch('icSave')}>
+		<Button disabled={invalid} on:click={() => dispatch('icSave')}>
 			{$i18n.tokens.import.text.add_the_token}
-		</ButtonPrimary>
+		</Button>
 	</ButtonGroup>
 </ContentWithToolbar>
