+++ src/frontend/src/icp/components/send/IcSendForm.svelte
@@ -7,8 +7,8 @@
 	import { icrcAccountIdentifierText } from '$icp/derived/ic.derived';
 	import type { IcAmountAssertionError } from '$icp/types/ic-send';
 	import SendSource from '$lib/components/send/SendSource.svelte';
+	import Button from '$lib/components/ui/Button.svelte';
 	import ButtonGroup from '$lib/components/ui/ButtonGroup.svelte';
-	import ButtonPrimary from '$lib/components/ui/ButtonPrimary.svelte';
 	import ContentWithToolbar from '$lib/components/ui/ContentWithToolbar.svelte';
 	import { balance } from '$lib/derived/balances.derived';
 	import { i18n } from '$lib/stores/i18n.store';
@@ -45,9 +45,9 @@
 
 		<ButtonGroup slot="toolbar">
 			<slot name="cancel" />
-			<ButtonPrimary disabled={invalid}>
+			<Button disabled={invalid}>
 				{$i18n.core.text.next}
-			</ButtonPrimary>
+			</Button>
 		</ButtonGroup>
 	</ContentWithToolbar>
 </form>
