+++ src/frontend/src/icp/components/send/IcSendReview.svelte
@@ -6,8 +6,8 @@
 	import { icrcAccountIdentifierText } from '$icp/derived/ic.derived';
 	import { isInvalidDestinationIc } from '$icp/utils/ic-send.utils';
 	import SendData from '$lib/components/send/SendData.svelte';
+	import Button from '$lib/components/ui/Button.svelte';
 	import ButtonGroup from '$lib/components/ui/ButtonGroup.svelte';
-	import ButtonPrimary from '$lib/components/ui/ButtonPrimary.svelte';
 	import ContentWithToolbar from '$lib/components/ui/ContentWithToolbar.svelte';
 	import { balance } from '$lib/derived/balances.derived';
 	import { tokenStandard } from '$lib/derived/token.derived';
@@ -49,8 +49,8 @@
 		<button class="secondary block flex-1" on:click={() => dispatch('icBack')}
 			>{$i18n.core.text.back}</button
 		>
-		<ButtonPrimary disabled={invalid} on:click={() => dispatch('icSend')}>
+		<Button disabled={invalid} on:click={() => dispatch('icSend')}>
 			{$i18n.send.text.send}
-		</ButtonPrimary>
+		</Button>
 	</ButtonGroup>
 </ContentWithToolbar>
