+++ src/frontend/src/icp/components/tokens/IcAddTokenReview.svelte
@@ -8,8 +8,8 @@
 		type ValidateTokenData
 	} from '$icp/services/ic-add-custom-tokens.service';
 	import AddTokenWarning from '$lib/components/tokens/AddTokenWarning.svelte';
+	import Button from '$lib/components/ui/Button.svelte';
 	import ButtonGroup from '$lib/components/ui/ButtonGroup.svelte';
-	import ButtonPrimary from '$lib/components/ui/ButtonPrimary.svelte';
 	import Card from '$lib/components/ui/Card.svelte';
 	import Logo from '$lib/components/ui/Logo.svelte';
 	import SkeletonCardWithoutAmount from '$lib/components/ui/SkeletonCardWithoutAmount.svelte';
@@ -99,9 +99,9 @@
 	<div in:fade>
 		<ButtonGroup>
 			<button class="secondary block flex-1" on:click={back}>{$i18n.core.text.back}</button>
-			<ButtonPrimary disabled={invalid} on:click={() => dispatch('icSave')}>
+			<Button disabled={invalid} on:click={() => dispatch('icSave')}>
 				{$i18n.tokens.import.text.add_the_token}
-			</ButtonPrimary>
+			</Button>
 		</ButtonGroup>
 	</div>
 {/if}
