+++ src/frontend/src/lib/components/manage/AddTokenByNetworkToolbar.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { createEventDispatcher } from 'svelte';
+	import Button from '$lib/components/ui/Button.svelte';
 	import ButtonGroup from '$lib/components/ui/ButtonGroup.svelte';
-	import ButtonPrimary from '$lib/components/ui/ButtonPrimary.svelte';
 	import { i18n } from '$lib/stores/i18n.store';
 
 	export let invalid: boolean;
@@ -13,7 +13,7 @@
 	<button type="button" class="secondary block flex-1" on:click={() => dispatch('icBack')}
 		>{$i18n.core.text.back}</button
 	>
-	<ButtonPrimary disabled={invalid}>
+	<Button disabled={invalid}>
 		{$i18n.core.text.next}
-	</ButtonPrimary>
+	</Button>
 </ButtonGroup>
