+++ src/frontend/src/lib/components/tokens/HideTokenReview.svelte
@@ -2,8 +2,8 @@
 	import { Html } from '@dfinity/gix-components';
 	import { nonNullish } from '@dfinity/utils';
 	import { createEventDispatcher } from 'svelte';
+	import Button from '$lib/components/ui/Button.svelte';
 	import ButtonGroup from '$lib/components/ui/ButtonGroup.svelte';
-	import ButtonPrimary from '$lib/components/ui/ButtonPrimary.svelte';
 	import ContentWithToolbar from '$lib/components/ui/ContentWithToolbar.svelte';
 	import Logo from '$lib/components/ui/Logo.svelte';
 	import { i18n } from '$lib/stores/i18n.store';
@@ -39,9 +39,9 @@
 		<button class="secondary block flex-1" on:click={() => dispatch('icCancel')}
 			>{$i18n.core.text.cancel}</button
 		>
-		<ButtonPrimary on:click={() => dispatch('icHide')}>
+		<Button on:click={() => dispatch('icHide')}>
 			{$i18n.tokens.hide.confirm}
-		</ButtonPrimary>
+		</Button>
 	</ButtonGroup>
 </ContentWithToolbar>
